requirements.txt
# Market data
requests>=2.31.0

# Scrapers
playwright>=1.40.0
twikit>=2.0.0
praw>=7.7.0

# MiroFish integration: uses `claude` CLI (OAuth), no SDK needed

# Utils
python-dotenv>=1.0.0
numpy>=1.24.0
